requirements.txt
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
colorama==0.4.6; platform_system == "Windows"
flask==3.0.3
itsdangerous==2.2.0
jinja2==3.1.4
markupsafe==2.1.5
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
tqdm==4.66.5
typing-extensions==4.12.2
urllib3==2.2.3
werkzeug==3.0.4